redefine functions
